+++ hub-app/src/main/java/datastore/hub_api/route/HubRoutes.java
@@ -1,5 +1,5 @@
 package datastore.hub_api.route;
-
+import datastore.hub_api.repository.HubRepository;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -143,21 +143,6 @@ public class HubRoutes {
                 .bodyToMono(String.class).block();
         }    
         return Response.noContent().build();
-    }
-
-    /**
-     * Converts Java group objects, whihc contian ids and urls, and converts them into a database table
-     */
-    @ApplicationScoped
-    public class HubRepository implements PanacheRepository<GroupURL> {
-        public Map<Long, URL> getGroupUrls() {
-            Map<Long, URL> IdMap = new HashMap<>();
-            //listAll lists all of the url Paneche entities
-            for (GroupURL group_url : listAll()) {
-                IdMap.put(group_url.id, group_url.url);
-            }
-            return IdMap;
-        }
     }
 }
